# Backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: CareFlow Orchestration API
  version: 0.1.0
servers:
  - url: http://localhost:8080
paths:
  /v1/patients:
    post:
      summary: Create patient reference
      requestBody:
        required: true
      responses:
        '201':
          description: Created
...
  /v1/journeys/{journeyId}:
    get:
      summary: Get care journey and timeline
      parameters:
        - name: journeyId
          in: path
          required: true
          schema: {type: string, format: uuid}
      responses:
        '200': {description: Journey}
        '404': {description: Not found}
...